// src/app/api/swagger/models/api-address-dto.ts
/* tslint:disable */
/* eslint-disable */
export interface ApiAddressDto {
  Area?: string | null;
  City?: string | null;
  Country?: string | null;
  CountryCode?: string | null;
  District?: string | null;
  FullAddress?: string | null;
  FullPostAddress?: string | null;
  Home?: string | null;
  Id?: number;
  KladrCodeArea?: string | null;
  KladrCodeCity?: string | null;
  KladrCodePopulatedLocality?: string | null;
  KladrCodeRegionRF?: string | null;
  KladrCodeStreet?: string | null;
  LocationAddress?: string | null;
  Office?: string | null;
  Okato?: string | null;
  PopulatedLocality?: string | null;
  PostIndex?: string | null;
  RegionRF?: string | null;
  Street?: string | null;
}